+++ mathmakerlib/__init__.py
@@ -25,7 +25,7 @@ from . import locale_patch, calculus, core, geometry, LaTeX
 __all__ = ['required', 'mmlib_setup', 'LaTeX', 'exceptions',
            'locale_patch', 'core', 'calculus', 'geometry']
 
-__version_info__ = (0, 6, 4)
+__version_info__ = (0, 7, 0)
 __dev__ = 0
 __patch_nb__ = 0
 __version__ = '.'.join(str(c) for c in __version_info__)